+++ lib/models/project.rb
@@ -12,6 +12,19 @@ module OpenBuildServiceAPI
       @name
     end
 
+    def meta(opts = {})
+      if !@cached_meta || @meta_reload
+        @cached_meta = @connection.send_request(:get, "/source/#{CGI.escape(@name)}/_meta").body
+        @meta_reload = false
+      end
+
+      if opts[:no_parse]
+        @cached_meta
+      else
+        Nokogiri::XML(@cached_meta)
+      end
+    end
+
     def delete!(message=nil)
       begin
         @connection.send_request(:delete, "/source/#{CGI.escape(@name)}", comment: message)
@@ -67,6 +80,7 @@ module OpenBuildServiceAPI
     def reload!
       @package_reload = true
       @public_key_reload = true
+      @meta_reload = true
     end
   end
 end