+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - Added `#rebuild!` method to `Package` model to allow rebuilding of packages
 - Added `#meta` method to `Project` model to obtain a project's meta config file
 - Added `#meta` method to `Package` model to obtain the meta definition of a package
+- Added `#reload!` method to `Package` model to reload the cached data of a package
 
 ## Version: 0.1.0 (alpha)