+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 - Added `#reload!` method to `Package` model to reload the cached data of a package
 - Added `#title` method to `Package` model to obtain the title of a package
 - Added `#description` method to `Package` model to obtain the description of a package
+- Improve output of `#inspect` for `Package` model instances
 
 ## Version: 0.1.0 (alpha)