+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## Version: 0.3.0 (alpha)
+
+- Added `#rebuild_failed!` method to `Package` model to rebuild only failed builds
+
 ## Version: 0.2.0 (alpha)
 
 - Added `#public_key` method to `Project` model to retrieve a project's public key